Cleanup main
